move useful scripts to pycolmap
