+++ pycolmap/pycolmap/utils/read_write_dense.py
