+++ pycolmap/pycolmap/utils/read_write_fused_vis.py
