+++ pycolmap/pycolmap/utils/read_write_model.py
